Improved remaining disk space calculation
Remaining disk space was not getting updated on item deletion. It will
now correctly reflect when an item is uploaded and when an item is
deleted it will now calculate the correct bytes in the background but
won't reflect automatically. Values will get updated after a refresh which
I think is good enough for now.